Consolidate TableDetail related containers
