+++ amundsen_application/static/js/components/TableDetail/DetailList/DetailListItem/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import moment from 'moment-timezone';
 
-import ColumnDescEditableText from '../../../../containers/TableDetail/ColumnDescEditableText';
+import ColumnDescEditableText from '../../ColumnDescEditableText';
 import { TableColumn } from '../../types';
 
 // TODO: Use css-modules instead of 'import'
