+++ amundsen_application/static/js/components/TableDetail/OwnerEditor/styles.scss
