+++ amundsen_application/static/js/components/TableDetail/OwnerEditor/types.ts
